requirements.txt
channels==4.2.2
daphne==4.1.2
Django==5.2
djangorestframework==3.16.0
fastapi==0.115.9
langchain==0.3.25
numpy==2.2.5
openai==1.78.1
pydantic==2.11.4
redis==6.1.0
requests==2.32.3
scikit-learn==1.6.1
sentence-transformers==4.1.0
torch==2.7.0
transformers==4.51.3
uvicorn==0.34.2